Enable copy paste for morse code
Text actions, like copy and paste, were only enabled
for text but not morse code. This was caused by a touch
listener that swallowed touch events when the text view
was in the "morse" mode.

Remove the touch event swallowing touch listener, and
a touch listener in the text state that just returned false.

Resolves #76